refactor(voip): drop dead code, type the socket seam, split files, remove comments
Cleanup pass over @zapo-js/voip:

- Dead code: remove udp-relay.ts (constructed but never connected — its
  hasConnection() branches were always false), the baileys-era
  processOfferNode/processAcceptNode/encodeNodeToBase64 signaling helpers,
  and the unused OPUS_DTX_FRAME/wasSendingSilence fields.
- Type the seam: replace `sock: any` in signaling/call-manager with the
  VoipSocket contract, which surfaced and fixed latent `string | undefined`
  bugs the `any` hid; tighten VoipSocket (query always returns a Promise,
  keys.get typed).
- Split long files: extract generateSecureSsrc -> ssrc.ts and the pure
  parseRelayFromAck -> relay-ack.ts out of call-manager.ts.
- Logging: route the manager's diagnostics through zapo's injected Logger
  (createNoopLogger default) instead of unconditional console.log; expose
  `logger` on createVoipManager options.
- Strip all comments from the package sources.

Builds CJS+ESM, typechecks, lints and passes unit + native-codec smoke tests.
