+++ packages/voip/src/encryption.ts
@@ -10,8 +10,6 @@ export async function derivePerJidSrtpKey(
     callKey: Buffer,
     deviceJid: string
 ): Promise<SrtpKeyingMaterial> {
-    // zapo's hkdf is sync with signature (ikm, salt, info, outLength) and an
-    // empty salt when null — matching baileys' hkdf(key, 46, { info }).
     const output = hkdf(callKey, null, textEncoder.encode(deviceJid), 46)
     return {
         masterKey: Buffer.from(output.subarray(0, 16)),
